Rename modules package to swerve
